requirements.txt
# Core
click==8.1.7
rich==13.7.0
pyyaml==6.0.1
watchdog==4.0.0

# AI
ollama==0.1.0

# Content Analysis
pdfplumber==0.10.0
python-docx==1.1.0
openpyxl==3.1.0
pytesseract==0.3.10
Pillow==10.0.0
eml-parser==1.17.0

# Testing
pytest==8.3.0
pytest-cov==6.0.0
